eb-flask/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
boto3==1.34.151
botocore==1.34.151
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
idna==3.7
ImageHash==4.3.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.0.1
pillow==10.4.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyWavelets==1.6.0
requests==2.32.3
s3transfer==0.10.2
scipy==1.13.1
six==1.16.0
urllib3==1.26.19
Werkzeug==3.0.3
zipp==3.19.2
